Extract IRI renumeration and script eval into their own modules
`inject-element.ts` was 397 lines, over half of it two self-contained
transforms on the loaded SVG. Both now live in their own modules and
`inject-element.ts` is 187 lines: resolve URL, load, transfer
attributes, delegate, swap into the DOM.

`eval-svg-scripts.ts` owns the `ranScripts` set that backs `evalScripts:
'once'`. `renumerate-svg-iri-elements.ts` owns the IRI element/property
table and the two reference-replacement helpers; those and the derived
set of referencing properties are now built once at module scope rather
than on every injection.

The exported names carry an `Svg` infix: `evalScripts` would shadow
`injectElement`'s parameter of that name, and `renumerateIriElements`
would differ from its `renumerateIRIElements` parameter only by case.
Test file names still track the option names.

No behaviour change, and no test edits. Confirmed by diffing the built
ESM bundle against the previous build: the moved statements are
identical, only relocated. Bundle 4.17 -> 4.22 kB gzipped, the
module-scope constants no longer being tree-shaken into the call site.
